extract Yelp data to GCS
This is the first attempt at extracting Yelp data with Yelp's limited daily API calls. It extracts and saves restaurant data based on matching the latitude and longitude of NYC open-source restaurants. It then is saved to Google Cloud Storage as an intended CSV file. We faced issues:

1. The restaurant's information was saved as individual CSV files instead of one CSV file with all restaurants
2. No guarantee that running the same script the following day would not extract the same restaurants
3. Risks a waste of Yelp API calls 
4. Difficult to compare the Yelp data to open-source because there was no same column to join the two datasets. Therefore proving difficult to clean and analyze Yelp's extracted data 

Our team reached out to an advisor on how we can improve our script for API yelp calls