Yelp API MongoDB
New code for connecting to MongoDB in order to store the Yelp data that was extracted and use it as a reference database. We switched databases for Yelp data because it is easier to store, has an index to look for specific restaurants, and resulted in using fuzzy matching. 

The script includes Yelp API extraction, accessing GCS for the NYC open source data to match restaurant longitude and latitude, access to MongoDB to store and compare Yelp calls, and appending the rating results to the open source data frame. If the restaurant data Yelp is looking to use for a call matches the data in MongoDB then it is skipped and added to a list of skips meanwhile if it is not found then it is extracted and stored in MongoDB. The appended data frame is then saved as a CSV.